// Frontend/client/src/api/models/PostDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PostDTO
 */
export interface PostDTO {
    /**
     * id of the post
     * @type {number}
     * @memberof PostDTO
     */
    id: number;
    /**
     * username of the user
     * @type {string}
     * @memberof PostDTO
     */
    username: string;
    /**
     * description of the post
     * @type {string}
     * @memberof PostDTO
     */
    description: string;
    /**
     * id of the image after using the pipeline
     * @type {number}
     * @memberof PostDTO
     */
    resultImageId: number;
    /**
     * id of the image before using the pipeline
     * @type {number}
     * @memberof PostDTO
     */
    originalImageId: number;
    /**
     * id of the pipeline
     * @type {number}
     * @memberof PostDTO
     */
    pipelineId: number;
    /**
     * Visibility of the post: 0 - private, 1 - friends, 2 - public, 10x where x is group ID for specific groups
     * @type {number}
     * @memberof PostDTO
     */
    visibility: number;
    /**
     * date of creation of the post
     * @type {Date}
     * @memberof PostDTO
     */
    createdAt: Date;
}

/**
 * Check if a given object implements the PostDTO interface.
 */
export function instanceOfPostDTO(value: object): value is PostDTO {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('username' in value) || value['username'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    if (!('resultImageId' in value) || value['resultImageId'] === undefined) return false;
    if (!('originalImageId' in value) || value['originalImageId'] === undefined) return false;
    if (!('pipelineId' in value) || value['pipelineId'] === undefined) return false;
    if (!('visibility' in value) || value['visibility'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    return true;
}

export function PostDTOFromJSON(json: any): PostDTO {
    return PostDTOFromJSONTyped(json, false);
}

export function PostDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): PostDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'username': json['username'],
        'description': json['description'],
        'resultImageId': json['result_image_id'],
        'originalImageId': json['original_image_id'],
        'pipelineId': json['pipeline_id'],
        'visibility': json['visibility'],
        'createdAt': (new Date(json['created_at'])),
    };
}

export function PostDTOToJSON(json: any): PostDTO {
    return PostDTOToJSONTyped(json, false);
}

export function PostDTOToJSONTyped(value?: PostDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'username': value['username'],
        'description': value['description'],
        'result_image_id': value['resultImageId'],
        'original_image_id': value['originalImageId'],
        'pipeline_id': value['pipelineId'],
        'visibility': value['visibility'],
        'created_at': ((value['createdAt']).toISOString()),
    };
}